Give focus to the question that was highlighted
A decision was made in 1806717 not to give any question focus if a field list was currently displayed. This was because focus was always given to the first question if it was an EditText-based question and that was confusing when a screen was jumped to to highlight a constraint violation. To avoid that confusion, this commit gives focus to the highlighted question which is generally what would be expected.