Move saving answers and rebuilding view to widgets
Rebuilding the view on activity result was introduced in 920b6a2. That commit says it fixes https://code.google.com/archive/p/opendatakit/issues/619 which looks unrelated to view refreshing and https://code.google.com/archive/p/opendatakit/issues/567. The form in 567 has been verified.

The combination of saving the answers for the screen and rebuilding the view in onActivityResult seemed to be a way to get around the fact that binary widgets did not update their UI to reflect a newly-selected value. This commit makes each binary widget refresh their own answer view and removes view rebuilding in onActivityResult and onResume.