Remove check for Aggregate 0.9 legacy servers (#3038)
The odds of servers of this type still running are low. If there are some, it would be good for the users to hit an issue and complain so they can be urged to upgrade away from ~8-year-old servers. The implementation was wrong anyway because it looked for a Location header rather than looking at the X-OpenRosa-Version header.