Refresh questions in intent groups after results are received
The intent group functionality used to rely on a full redraw of the current question container view. Now, each question gets refreshed after its value changes in the model. It is safe to cast any widget in an intent group to StringWidget because only text, integer and decimal types are supported and all corresponding widgets are descendants of StringWidget.