Resume location capture after activity killed by system
In that case, the form doesn't need to be loaded so we don't get a form loaded event. If the form audits location, we should go straight to requesting locations or stopping if preconditions aren't met.

Removes the savedFormStart flag that was used to detect this case previously.

There is now one known limitation: if FormEntryActivity is killed while the hierarchy view is being shown after a saved form start, nothing related to background location will be displayed to the user or logged to the audit log when the activity does become visible.