Add material theme attributes
We need to do this to use components like the MaterialButton. I've kept the colors
as identical as possible so there should be no visual changes. Moving to a Material
Theme would have been the alternative approach but this would have introduced changes
to older components in the app like the normal Buttons.