Fix opening instance when moving backwards is disabled
Prior to 0a6bb3d, it was possible to open instances for viewing when moving backwards was disabled but 0a6bb3d broke that by moving the mode detection (view sent vs edit saved) into an else block. This logic needs some rethinking but I've left a TODO for now because it's not related to the form map view.