Remove SNI test
This looks like it had been designed to drive out a fix in the old
ODK Apache HTTP client. We could add a test to our OpenRosaHttpInterface
but it feels like overkill as SNI is supported by Java's own HttpURLConnection.
I'd be very surprised if any http clients we use in the future have do not
include SNI support given that's the case.