Don't set the map field before initialization
When onResume called updateInstanceGeometry, the map wasn't null so there was an attempt to map features even though the map fragment was not guaranteed to be fully initialized. We need the geometry update in onResume for the case where a user triggers form filling from this activity and then returns to it having added a new feature.