Track launches of forms with background location actions
This was a very-requested feature and it would be good to know whether it's actually in use or whether we need to do more messaging around it. Note that this slightly changes the behavior of the form definition hash: if the form controller is null, the hash will be an empty string rather than a hash of the empty string. We don't want to count that case because it wasn't actually a successful form launchso this will make it easier to identify. If it happens often, we have a problem. Most calls including this one should already have checked for the null form controller case.