Always call onClientStop from stop, never from onConnectionSuspended
According to the docs on LocationClientListener.onClientStop, it should always be called when stop() is called. This also matches the AndroidLocationClient implementation. onConnectionSuspended is called when the connection to Google Play Services is lost because, for example, the user stops it. According to the original author of the API, a reconnection attempt will be made: https://stackoverflow.com/a/26147518/137744. I don't think it's appropriate to consider the LocationClient stopped at that point.