Change service start type
This prevents the audio recorder service from being restarted if
the app process is restarted. If this happens the service is started
with a `null` intent and so ends up in a weird state where it shows
a notification (as it was previously a foreground service) but
doesn't do anything and can't be dismissed.