Remove ReferenceManager-related tests
These still capture the intended end state but there's no interaction with ReferenceManager in form metadata parsing anymore so they make sense to remove. It's also the case that if a form was just opened and then is updated, that form's media files would resolve during and after the medata parse because the ReferenceManager is only reset at the beginning of a form filling session.