Correct testshared dependencies
`testshared` should only ever really be built during debug builds
(as it will only ever be a `testImplementation` dep) but the
Android plugin's assemble tasks (when used on the app module)
builds every module in the project annoyingly. This means that
`debugImplementation` deps will be missing for release builds of
test only dependencies.