Remove broken onSaveInstanceState
It looks like onSaveInstanceState was saving a path to `outState`
(correct), but that `onCreate` was then trying to read that state
from the intent extras (incorrect). Testing the activity with
"Don't keep activities" demonstrated that the image would be lost
due to this.

Additionally onSaveInstanceState was compressing a bitmap and
saving a file on the UI thread which could lead to ANRs. This whole
process would need to be reworked anyway, so I felt like removing
the "dead" code made more sense than attempting to fix it while
working on ANR fixes.